pyproject.toml
[project]
name = "class-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "ruff>=0.9.9",
    "streamlit>=1.43.0",
]
